% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{ddt data set}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{LENGTH}{length in cm}
  \item{WEIGHT}{weight in grams}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ddt
}
\description{
A data set collected by the US government in a Tenessee river
}
\keyword{datasets}
