% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iter is the number of iterations to be run}

\item{n}{is the number of trials}

\item{p}{is the probability of each possibility}
}
\value{
Makes a plot
}
\description{
mybin
}
\examples{
\dontrun{mybin(iter = 100, n = 10, p = 0.5)}
}
