% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybootNew.R
\name{mybootNew}
\alias{mybootNew}
\title{mybootNew}
\usage{
mybootNew(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{the sample to be used}

\item{fun}{the type of statistic to be measured}

\item{alpha}{confidence interval}

\item{...}{}
}
\value{
a vector containing a list of various stats
}
\description{
mybootNew
}
