% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDDT.R
\name{myDDT}
\alias{myDDT}
\title{myDDT}
\usage{
myDDT(ddt, Species)
}
\arguments{
\item{ddt}{the data set to be used}

\item{Species}{}
}
\value{
A plot
}
\description{
myDDT
}
\examples{
\dontrun{myDDT(ddt, Species = "CCATFISH") }
}
